pyproject.toml
[project]
name = "starting-codebase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb==1.0.15",
    "zhipuai>=2.1.0",
    "sentence-transformers==5.0.0",
    "fastapi==0.116.1",
    "uvicorn==0.35.0",
    "python-multipart==0.0.20",
    "python-dotenv==1.1.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "requests>=2.32.4",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "flake8-docstrings>=1.7.0",
    "pylint>=3.0.0",
]

[tool.black]
line-length = 100
target-version = ['py313']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.ruff_cache
    | \.venv
    | \.eggs
    | \.eggs
    | dist
    | build
    | venv
    | env
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pylint.messages_control]
disable = [
    "C0111",  # 禁用缺少 docstrings 的警告
    "C0103",  # 允许短变量名
    "R0903",  # 允许少方法类
]
max-line-length = 100

[tool.pylint.format]
max-line-length = 100
